Adds several chat commands like the ability to message other players or reply to them, as well as the shout command to talk between worlds
Chat in SmashServers as well as game announcements is now limited to the server itself, and players leaving or joining your server is now announced by chat message

Adds the all in one world command for editing base maps, loading exact worlds, creating base maps, teleporting to, and listing worlds

Adds items that allow you to browse servers, select kits, teleport to the hub, and vote for maps

Villager values have been adjusted to be more accurate

Doublejump now checks if the kit is active rather than if the game hotbar is equipped, fixing a bug where you could toggle flight while the game was starting

Fixes a bug where when disguises were removed the player was still invisible unless they changed worlds, to fix this I call the bukkit function for hiding a player before showing them, since showing them does nothing unless bukkit was tracking them as being hidden

Re-adds Kit Podiums

Servers now have a different material in the server list, by default it is an emerald block, it is a gold block if the server is currently playing, and it becomes a redstone block if the server is full

Adds kit pre-selection, this makes it so you no longer equip a kit in a lobby until you are past the voting phase, which lets you keep the item in your inventory that allows you to vote in the first place

Folders for gamemode maps now actually check if they have a region directory before instantly assuming they are a world folder

Allowed kits now clear properly before they re-add themselves

Adds improved map parsing as well as a world command to properly convert old maps to the new system. Instead of parsing the entirely world inefficiently, we put item frames on top of the blocks that we want to parse. This means instead of searching hundreds of thousands of blocks we simply use world.getEntities() which allows us to get the points exact positions. This means when we load new game maps the loading time doesn't even cause a single tick of visible lag on the server at all.

Tries to fix game worlds from autosaving which means the next time they are loaded all of their point data is gone

Fixes a bug where you didn't get your items back after being teleported to the lobby because you had your kit unequipped after teleporting which would clear your items

Lowers Fish Flurry, Slime Rocket, and Cub Tackle projectile hitbox sizes to be in-line with Block Toss since they should be the same size

Fixes a bug where players could be damaged when they didn't have a server which meant they would take falling damage in the main world