Moves base mode kits and abilities into their own packages so we can organize them by kits and abilities that are different enough and part of seperate modes
You can no longer equip the hub kit in the SmashServer lobbies in debug mode

Adds a new boss bar manager for assigning boss bars to players

The damage manager no longer condenses records again since it was causing issues

Getting random respawn points is now the job of the gamemode

Gamemodes now have an update function that runs before the smashserver updates

Void damage is now directly assigned to the player that caused it

Adds a utility function to get an armor amount that gives you the effective health passed

Adds a new boss kit, the Elder Guardian with it's whirlpool shotgun (ink shotgun + whirlpool axe)

Adds a new attribute that stops true damage besides void damage with an optional damage multiplier

Adds a new attribute that stops void deaths and lets you choose how much damage you take instead and if it ignores armor

Adds a new boss gamemode, every player in the game matches up against a single other player who is chosen to be the boss. Health scales based on the amount of active players in the game and placements are based on how much damage was dealt to the boss by that individual player. The boss will always respawn in the maps center currently.