Fixes wither skull sending cooldown messages when it shouldnt
Fixes border lightning spamming

Melee is handled a little differently, instead of redirecting packets that attack mobs directly to the player, a new packet is made on the players connection to melee the opposing player. This is how mineplex did it and may or may not make melee feel more accurate.